+++ src/briefy/leica/models/customer/__init__.py
@@ -76,7 +76,7 @@ class Customer(TaxInfo, mixins.PolaroidMixin, mixins.CustomerBriefyRoles,
     __summary_attributes__ = [
         'id', 'slug', 'title', 'description', 'created_at', 'updated_at', 'state', 'external_id'
     ]
-
+    __summary_attributes_relations__ = ['billing_contact', 'business_contact', 'addresses']
     __listing_attributes__ = __summary_attributes__
 
     __colanderalchemy_config__ = {'excludes': [
@@ -119,7 +119,7 @@ class Customer(TaxInfo, mixins.PolaroidMixin, mixins.CustomerBriefyRoles,
 
     addresses = orm.relationship(
         'CustomerBillingAddress',
-        backref=orm.backref('customer', lazy='joined'),
+        backref=orm.backref('customer'),
         lazy='dynamic',
     )
     """List of Billing Addresses for a Customer
@@ -196,20 +196,14 @@ class Customer(TaxInfo, mixins.PolaroidMixin, mixins.CustomerBriefyRoles,
         Used to serialize this object within a parent object serialization.
         :returns: Dictionary with fields and values used by this Class
         """
-        billing_contact = self.billing_contact
-        business_contact = self.business_contact
         data = super().to_listing_dict()
         data = self._apply_actors_info(data)
-        data['billing_contact'] = billing_contact.to_summary_dict() if billing_contact else None
-        data['business_contact'] = business_contact.to_summary_dict() if business_contact else None
         return data
 
     def to_dict(self):
         """Return a dict representation of this object."""
         data = super().to_dict()
         data['slug'] = self.slug
-        addresses = [address.to_dict(excludes='customer') for address in self.addresses.all()]
-        data.update(addresses=addresses)
         add_user_info_to_state_history(self.state_history)
         # Apply actor information to data
         data = self._apply_actors_info(data)