+++ src/briefy/leica/models/professional/__init__.py
@@ -63,6 +63,8 @@ class Professional(ProfessionalMixin, Base):
         'main_email', 'main_mobile', 'type', 'photo_path'
     ]
 
+    __summary_attributes_relations__ = ['links', 'main_location']
+
     __listing_attributes__ = __summary_attributes__
 
     __colanderalchemy_config__ = {
@@ -111,7 +113,7 @@ class Professional(ProfessionalMixin, Base):
     # Assets
     assets = orm.relationship(
         'Asset',
-        backref=orm.backref('professional', lazy='joined'),
+        backref=orm.backref('professional'),
         lazy='dynamic'
     )
 
@@ -126,7 +128,7 @@ class Professional(ProfessionalMixin, Base):
     locations = orm.relationship(
         'WorkingLocation',
         backref=orm.backref(
-            'professional', lazy='joined'
+            'professional'
         ),
         cascade='all, delete-orphan',
     )
@@ -161,45 +163,9 @@ class Professional(ProfessionalMixin, Base):
             args['polymorphic_on'] = cls.type
         return args
 
-    def _summarize_relationships(self) -> dict:
-        """Summarize relationship information.
-
-        :return: Dictionary with summarized info for relationships.
-        """
-        data = {}
-        links = self.links
-        main_location = self.main_location
-        to_summarize = [
-            ('links', links),
-        ]
-
-        if main_location:
-            to_summarize.append(
-                ('main_location', main_location)
-            )
-
-        for k, obj in to_summarize:
-            if isinstance(obj, Base):
-                serialized = obj.to_summary_dict() if obj else None
-            else:
-                serialized = [o.to_summary_dict() for o in obj]
-            data[k] = serialized
-        return data
-
-    def to_listing_dict(self) -> dict:
-        """Return a summarized version of the dict representation of this Class.
-
-        Used to serialize this object within a parent object serialization.
-        :returns: Dictionary with fields and values used by this Class
-        """
-        data = super().to_listing_dict()
-        data.update(self._summarize_relationships())
-        return data
-
     def to_dict(self):
         """Return a dict representation of this object."""
         data = super().to_dict()
-        data.update(self._summarize_relationships())
 
         # Workflow history
         add_user_info_to_state_history(self.state_history)